Update install.md
Added Ensime pinning to melpa-stable that is compatible with `:ensure t` and `(setq use-package-always-ensure t)`.  Note that package-pinned-packages was introduced in Emacs 24.4.  This variable is also used inside of use-package for its `:pin` implementation, so the Emacs version requirements haven't changed for this to work properly.  The key to making this work was to make sure `(package-refresh-contents)` is called after `package-pinned-packages` is updated.  You can either set `package-pinned-packages` before the first call to `(package-refresh-contents)` in init.el, or you can use the more complicated but self contained logic I've included in this PR.